requirements.txt
asgiref==3.4.1
Django==4.0
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
PyJWT==2.3.0
sqlparse==0.4.2
